Add --tmp-dir option to cewe2pdf
It allows people to manually specify a directory for .mcfx file extraction.
Files in this directory are not deleted when cewe2pdf ends.

When cewe2pdf is started the next time, the files are re-used
instead of being extracted again.
This makes it much faster to develop on cewe2pdf because files don't
have to be extracted again and again.

Files that have been updated in the .mcfx file get unpacked
and overwrite their older sibling in the tmp directory.

Related: https://github.com/bash0/cewe2pdf/issues/119