Add support for flipping cliparts horizontally and vertically
The "mirror" area attribute with values "x", "y" and "both" determines
if a clipart shall be flipped horizontally and/or vertically.

At first I tried to use `pdf.scale(1, -1)` to flip the clipart
via PDF means, but that led to positioning problems.
Flipping the image itself does not exhibit this problem.

The mirror="x" value means that the image should be flipped vertically,
which is .. a bit counter-intuitive.